rename workflow template file and rename its imported var to full caps
