+++ frontend/javascripts/oxalis/view/action-bar/default-predict-workflow-template.ts
